Split feed view controller delegates into multiple files
