+++ Secretly/ViewControllers/FeedCollectionViewController+UICollectionViewDataSourcePrefetching.swift
@@ -1,0 +1,16 @@
+//
+//  FeedCollectionViewController+UICollectionViewDataSourcePrefetching.swift
+//  Secretly
+//
+//  Created by Luis Ezcurdia on 12/06/21.
+//  Copyright © 2021 3zcurdia. All rights reserved.
+//
+
+import UIKit
+
+extension FeedCollectionViewController: UICollectionViewDataSourcePrefetching {
+    func collectionView(_ collectionView: UICollectionView, prefetchItemsAt indexPaths: [IndexPath]) {
+        guard let indexPath = indexPaths.last else { return }
+        print("================\(indexPath.row)=================")
+    }
+}
